Cleanup After Phpstan Upgrade
After Phpstan 1.6.3 upgrade, clean up some new problems that will show up if it runs under Php 8+.

+++ src/PhpSpreadsheet/Reader/Slk.php
@@ -450,7 +450,7 @@ class Slk extends BaseReader
 
                     break;
                 case 'M':
-                    $formatArray['font']['size'] = substr($rowDatum, 1) / 20;
+                    $formatArray['font']['size'] = ((float) substr($rowDatum, 1)) / 20;
 
                     break;
                 case 'L':
